added unit logik for start rating and changed schemas

+++ app/lib/actions.js
@@ -7,7 +7,7 @@ import { Unit } from "./models/Unit";
 import { connectToDB } from "./utils";
 import bcrypt from "bcrypt";
 import { signIn } from "../auth";
-import Rating from "./models/Rating";
+import { Rating } from "./models/Rating";
 
 /* ------------------------ User ---------------------------------------- */
 
@@ -20,17 +20,23 @@ export const addUser = async (formData) => {
     password,
     phone,
     position,
-    unit,
+    unit: unitId,
     role,
     isActive,
   } = Object.fromEntries(formData);
 
   try {
-    connectToDB();
+    await connectToDB();
 
     const salt = await bcrypt.genSalt(10);
     const hashedPassword = await bcrypt.hash(password, salt);
+
+    // Finde die höchste vorhandene userId und erhöhe sie um 1
+    const highestUser = await User.findOne().sort({ userId: -1 }).limit(1);
+    const userId = highestUser ? highestUser.userId + 1 : 1;
+
     const newUser = new User({
+      userId,
       firstname,
       lastname,
       dob,
@@ -38,11 +44,22 @@ export const addUser = async (formData) => {
       password: hashedPassword,
       phone,
       position,
-      unit,
+      unit: unitId,
       role,
       isActive,
     });
     await newUser.save();
+
+    const unit = await Unit.findOne({ unitId });
+    if (unit) {
+      // If the new user is the manager, avoid double-counting
+      if (unit.manager !== userId) {
+        unit.employees = unit.employees + 1;
+      } else if (unit.employees === 0) {
+        unit.employees = 1;
+      }
+      await unit.save();
+    }
   } catch (error) {
     console.log(error);
     throw new Error("Failed to create User!");
@@ -59,7 +76,6 @@ export const updateUser = async (formData) => {
     lastname,
     dob,
     email,
-    password,
     phone,
     position,
     unit,
@@ -67,17 +83,18 @@ export const updateUser = async (formData) => {
     isActive,
   } = Object.fromEntries(formData);
 
-  const salt = await bcrypt.genSalt(10);
-  const hashedPassword = await bcrypt.hash(password, salt);
-
   try {
     connectToDB();
+
+    const user = await User.findById(id);
+    const oldUnitId = user.unit;
+
+    // Aktualisiere die Benutzerdaten
     const updateFields = {
       firstname,
       lastname,
       dob,
       email,
-      password: hashedPassword,
       phone,
       position,
       unit,
@@ -86,10 +103,27 @@ export const updateUser = async (formData) => {
     };
     Object.keys(updateFields).forEach(
       (key) =>
-        (updateFields[key] === "" || undefined) && delete updateFields[key]
+        (updateFields[key] === "" || updateFields[key] === undefined) &&
+        delete updateFields[key]
     );
 
     await User.findByIdAndUpdate(id, updateFields);
+
+    // Wenn die Unit geändert wurde, aktualisiere die Mitarbeiteranzahlen
+    if (oldUnitId !== unit) {
+      const oldUnit = await Unit.findOne({ unitId: oldUnitId });
+      const newUnit = await Unit.findOne({ unitId: unit });
+
+      if (oldUnit) {
+        oldUnit.employees -= 1;
+        await oldUnit.save();
+      }
+
+      if (newUnit) {
+        newUnit.employees += 1;
+        await newUnit.save();
+      }
+    }
   } catch (error) {
     console.log(error);
     throw new Error("Failed to update User!");
@@ -103,7 +137,25 @@ export const deleteUser = async (formData) => {
   const { id } = Object.fromEntries(formData);
 
   try {
-    connectToDB();
+    await connectToDB();
+
+    const user = await User.findById(id);
+    if (!user) {
+      throw new Error("User not found");
+    }
+
+    const unit = await Unit.findOne({ unitId: user.unit });
+    if (unit) {
+      // If the deleted user is the manager, update the manager field and avoid double-counting
+      if (unit.manager === user.userId) {
+        unit.manager = null;
+        unit.employees = unit.employees - 1;
+      } else if (unit.employees > 0) {
+        unit.employees = unit.employees - 1;
+      }
+      await unit.save();
+    }
+
     await User.findByIdAndDelete(id);
   } catch (error) {
     console.log(error);
@@ -116,15 +168,33 @@ export const deleteUser = async (formData) => {
 /* ------------------------ Unit ---------------------------------------- */
 
 export const addUnit = async (formData) => {
-  const { unitId, name, location, area } = Object.fromEntries(formData);
+  const {
+    name,
+    location,
+    area,
+    manager,
+    contactPerson,
+    description,
+    parentUnit,
+  } = Object.fromEntries(formData);
 
   try {
-    connectToDB();
+    await connectToDB();
+
+    // Finde die höchste vorhandene unitId und erhöhe sie um 1
+    const highestUnit = await Unit.findOne().sort({ unitId: -1 }).limit(1);
+    const unitId = highestUnit ? highestUnit.unitId + 1 : 1;
+
     const newUnit = new Unit({
       unitId,
       name,
       location,
       area,
+      manager,
+      employees: manager ? 1 : 0,
+      contactPerson,
+      description,
+      parentUnit,
     });
     await newUnit.save();
   } catch (error) {
@@ -137,20 +207,52 @@ export const addUnit = async (formData) => {
 };
 
 export const updateUnit = async (formData) => {
-  const { id, unitId, name, location, area } = Object.fromEntries(formData);
+  const {
+    id,
+    name,
+    location,
+    area,
+    manager,
+    contactPerson,
+    description,
+    parentUnit,
+  } = Object.fromEntries(formData);
 
   try {
-    connectToDB();
+    await connectToDB();
+
+    // Fetch the current unit to check the current manager status
+    const currentUnit = await Unit.findById(id);
+    if (!currentUnit) {
+      throw new Error("Unit not found");
+    }
+
     const updateFields = {
-      unitId,
       name,
       location,
       area,
+      manager,
+      contactPerson,
+      description,
+      parentUnit,
     };
-    Object.keys(updateFields).forEach(
-      (key) =>
-        (updateFields[key] === "" || undefined) && delete updateFields[key]
-    );
+
+    // Update employees count if manager is added or removed
+    if (manager && !currentUnit.manager) {
+      // Check if the manager is the only person in the unit
+      if (currentUnit.employees === 0) {
+        updateFields.employees = 1;
+      } else {
+        updateFields.employees = currentUnit.employees + 1;
+      }
+    } else if (!manager && currentUnit.manager) {
+      updateFields.employees = currentUnit.employees - 1;
+    }
+
+    // Prevent double-counting if manager is also an employee
+    if (manager && currentUnit.manager === manager) {
+      updateFields.employees = currentUnit.employees;
+    }
 
     await Unit.findByIdAndUpdate(id, updateFields);
   } catch (error) {
@@ -166,7 +268,17 @@ export const deleteUnit = async (formData) => {
   const { id } = Object.fromEntries(formData);
 
   try {
-    connectToDB();
+    await connectToDB();
+
+    const unit = await Unit.findById(id);
+    if (!unit) {
+      throw new Error("Unit not found");
+    }
+
+    // Set unit ID to null for users associated with this unit
+    await User.updateMany({ unit: unit.unitId }, { $set: { unit: null } });
+
+    // Delete the unit
     await Unit.findByIdAndDelete(id);
   } catch (error) {
     console.log(error);
@@ -179,7 +291,8 @@ export const deleteUnit = async (formData) => {
 /* ------------------------ Rating ---------------------------------------- */
 
 export const addRating = async (formData) => {
-  const user_id = formData.get("user_id");
+  const userId = formData.get("userId");
+  const ratingUserId = formData.get("ratingUserId");
 
   const social = {};
   formData.forEach((value, key) => {
@@ -210,31 +323,39 @@ export const addRating = async (formData) => {
   try {
     connectToDB();
 
-    let existingRating = await Rating.findOne({ user_id });
+    // Find existing rating entry for the user
+    let existingRating = await Rating.findOne({ userId });
 
+    // If no existing entry, create a new one
     if (!existingRating) {
-      // Falls kein Rating vorhanden ist, ein neues erstellen
       existingRating = new Rating({
-        user_id,
-        ratings: [], // Initialisierung von ratings
+        userId,
+        ratings: [],
       });
-    } else if (!existingRating.rating) {
-      // Falls das vorhandene Rating keine Bewertungen hat, ein leeres Array initialisieren
-      existingRating.rating = [];
     }
 
+    // Find the highest existing ratingId in the entire database
+    const maxRatingId = await Rating.aggregate([
+      { $unwind: "$ratings" },
+      { $group: { _id: null, maxRatingId: { $max: "$ratings.ratingId" } } },
+    ]);
+
+    const newRatingId =
+      maxRatingId.length > 0 ? maxRatingId[0].maxRatingId + 1 : 1;
+
     const newRating = {
-      createdAt: new Date(),
-      lastUpdated: new Date(),
+      ratingId: newRatingId,
+      ratingUserId,
       social,
       kpi,
       additionalCriteria,
       totalScore,
     };
-    console.log(existingRating);
-    console.log(existingRating.rating);
-    existingRating.rating.push(newRating); // Hinzufügen des neuen Ratings zum Array
 
+    // Add the new rating to the existing ratings array
+    existingRating.ratings.push(newRating);
+
+    // Save the updated rating entry
     await existingRating.save();
   } catch (error) {
     console.log("Error creating rating:", error);
@@ -242,8 +363,8 @@ export const addRating = async (formData) => {
   }
 
   revalidatePath("/dashboard/ratings");
+  redirect("/dashboard/ratings");
 };
-
 /* ------------------------ Authentication ---------------------------------------- */
 
 export const authenticate = async (prevState, formData) => {
